Added Functionality
Added a filter to check for unallowed characters and names for the campaign names since some file systems don't like some stuff (cough) windows (cough).
Added the building blocks for storing campaign data.
Added the foundation of the loading of campaigns in main to the renderer.
Moved where the help button is.
Also fixed some css that was bothering me.